fix: align MTP draft models with target topology
